requirements.txt
# Bleeding edge Django
django==1.9.2

# Configuration
django-environ==0.4.0
django-extensions==1.6.1

# Models
django-model-utils==2.4